[SCA] SCA-based code improvements.

+++ jopa-impl/src/main/java/cz/cvut/kbss/jopa/query/soql/RdfContainerTriplePatternEnhancer.java
@@ -12,8 +12,8 @@ class RdfContainerTriplePatternEnhancer extends TriplePatternEnhancer {
     @Override
     List<String> getTriplePatterns(String subject, String predicate, String object) {
         // Make variable names dependent on predicate to avoid collisions. This is not bulletproof but should be good enough.
-        // Cannot use just hashCode because it may be negative which makes the variable name invalid in SPARQL
-        final int predicateHash = Math.abs(predicate.hashCode());
+        // Cannot use just hashCode because it may be negative, making the variable name invalid in SPARQL
+        final long predicateHash = Math.abs(predicate.hashCode());  // Widen to long as abs of hashcode may be negative
         final String containerVariable = "?rdfContainer" + predicateHash;
         final String hasElementVariable = "?hasElement" + predicateHash;
         return List.of(
